+++ src/components/DataManagerApp/DataImportView/GoogleSheetsImportView/index.tsx
@@ -4,8 +4,8 @@ import { invariant } from "@tanstack/react-router";
 import { useState } from "react";
 import { z } from "zod";
 import { APIClient } from "@/clients/APIClient";
-import { DatasetClient } from "@/clients/datsets/DatasetClient";
-import { DatasetRawDataClient } from "@/clients/datsets/DatasetRawDataClient";
+import { DatasetClient } from "@/clients/datasets/DatasetClient";
+import { DatasetRawDataClient } from "@/clients/datasets/DatasetRawDataClient";
 import { useGooglePicker } from "@/hooks/ui/useGooglePicker";
 import { useCurrentUserProfile } from "@/hooks/users/useCurrentUserProfile";
 import { useCurrentWorkspace } from "@/hooks/workspaces/useCurrentWorkspace";
