+++ src/components/DataManagerApp/DataImportView/ManualUploadView/index.tsx
@@ -3,8 +3,8 @@ import { Dropzone, FileWithPath } from "@mantine/dropzone";
 import { IconPhoto, IconUpload, IconX } from "@tabler/icons-react";
 import { useQueryClient } from "@tanstack/react-query";
 import { invariant } from "@tanstack/react-router";
-import { DatasetClient } from "@/clients/datsets/DatasetClient";
-import { DatasetRawDataClient } from "@/clients/datsets/DatasetRawDataClient";
+import { DatasetClient } from "@/clients/datasets/DatasetClient";
+import { DatasetRawDataClient } from "@/clients/datasets/DatasetRawDataClient";
 import { useCurrentUserProfile } from "@/hooks/users/useCurrentUserProfile";
 import { useCurrentWorkspace } from "@/hooks/workspaces/useCurrentWorkspace";
 import { MIMEType } from "@/lib/types/common";
