+++ src/components/DataManagerApp/DatasetMetaView/index.tsx
@@ -16,9 +16,9 @@ import { notifications } from "@mantine/notifications";
 import { IconPencil, IconX } from "@tabler/icons-react";
 import { useNavigate } from "@tanstack/react-router";
 import { useMemo, useState } from "react";
-import { DatasetClient } from "@/clients/datsets/DatasetClient";
-import { DatasetColumnClient } from "@/clients/datsets/DatasetColumnClient";
-import { DatasetRawDataClient } from "@/clients/datsets/DatasetRawDataClient";
+import { DatasetClient } from "@/clients/datasets/DatasetClient";
+import { DatasetColumnClient } from "@/clients/datasets/DatasetColumnClient";
+import { DatasetRawDataClient } from "@/clients/datasets/DatasetRawDataClient";
 import { AppConfig } from "@/config/AppConfig";
 import { AppLinks } from "@/config/AppLinks";
 import { useCurrentWorkspace } from "@/hooks/workspaces/useCurrentWorkspace";
