+++ src/components/DataManagerApp/index.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, MantineTheme } from "@mantine/core";
 import { Outlet } from "@tanstack/react-router";
-import { DatasetClient } from "@/clients/datsets/DatasetClient";
+import { DatasetClient } from "@/clients/datasets/DatasetClient";
 import { useCurrentWorkspace } from "@/hooks/workspaces/useCurrentWorkspace";
 import { where } from "@/lib/utils/filters/filterBuilders";
 import { DatasetNavbar } from "./DatasetNavbar";
