+++ src/components/EntityDesignerApp/EntityConfigMetaView/useHydratedEntityConfig.ts
@@ -1,5 +1,5 @@
 import { useMemo } from "react";
-import { DatasetClient } from "@/clients/datsets/DatasetClient";
+import { DatasetClient } from "@/clients/datasets/DatasetClient";
 import { where } from "@/lib/utils/filters/filterBuilders";
 import { DatasetId } from "@/models/datasets/Dataset";
 import { EntityFieldConfigClient } from "@/models/EntityConfig/EntityFieldConfig/EntityFieldConfigClient";
